feat: migrate the content store from SQLite to Neon Postgres
Consolidates onto the database provider already in use rather than adding
Turso alongside it. Done now because the coupling only grows: every page
still to build (/film, /playbook, /positions) would add more
SQLite-flavoured code to convert.

  10 tables      sqliteTable -> pgTable
  4 primary keys integer({autoIncrement}) -> serial
  6 columns      text({mode:"json"}) -> jsonb
  4 columns      integer({mode:"boolean"}) -> boolean
  1 raw query    published = 1 -> published = true
  sync.ts        SQLiteTable -> PgTable
  migrations     regenerated in the postgres dialect

Driver is neon-serverless over WebSockets, NOT neon-http, and that is not
incidental. neon-http's transaction() throws outright ("No transactions
support in neon-http driver"), and lib/admin/db-repo.ts wraps tag
replacement in db.transaction() because a code review found that a partial
failure there destroys a video's tags. Choosing the obvious driver would
have silently dropped that safety property at runtime. The reason is
recorded in client.ts so nobody simplifies it back.

Two dialect differences that only surfaced by running it:

  Postgres returns count() as bigint, which node-postgres hands back as a
  STRING to avoid precision loss. sql<number> is a compile-time lie, so
  the players index rendered "51 players · 08644444433332" -- string
  concatenation, not a sum. All 11 aggregates now cast ::int.

  There is no local-file fallback any more. getDb() previously defaulted
  to file:.data/coach-hayes.db so local dev needed no configuration;
  DATABASE_URL is now required everywhere, including CI builds, since
  generateStaticParams reads the database to emit the 51 player pages.
  The db:* and ingest scripts gained --env-file=.env accordingly.

Verified against a real Neon instance (Postgres 18.6): ingest produced
numbers identical to SQLite -- 455 videos, 285 published, 170 in review,
523 player links, 83 players with film -- the build emitted the same 51
static pages, and confirming a video in the admin wrote through
(published 285 -> 286, review 170 -> 169) with published reading back as
a real boolean and counts as numbers.

+++ lib/admin/db-repo.ts
@@ -205,7 +205,7 @@ export const dbRepo: AdminRepository = {
   async getSyncStatus(): Promise<SyncStatus> {
     const stored = await readMeta<SyncStatus>("sync");
     if (stored) return stored;
-    const [{ n }] = await getDb().select({ n: sql<number>`count(*)` }).from(videos);
+    const [{ n }] = await getDb().select({ n: sql<number>`count(*)::int` }).from(videos);
     return {
       lastSyncAt: null,
       state: n > 0 ? "ok" : "never-run",
@@ -218,10 +218,10 @@ export const dbRepo: AdminRepository = {
   async getPublishedCounts(): Promise<PublishedCounts> {
     const db = getDb();
     const [[v], [p], [c], [t]] = await Promise.all([
-      db.select({ n: sql<number>`count(*)` }).from(videos).where(eq(videos.published, true)),
-      db.select({ n: sql<number>`count(distinct ${videoPlayers.playerId})` }).from(videoPlayers),
-      db.select({ n: sql<number>`count(*)` }).from(concepts),
-      db.select({ n: sql<number>`count(distinct ${videoTopics.topic})` }).from(videoTopics),
+      db.select({ n: sql<number>`count(*)::int` }).from(videos).where(eq(videos.published, true)),
+      db.select({ n: sql<number>`count(distinct ${videoPlayers.playerId})::int` }).from(videoPlayers),
+      db.select({ n: sql<number>`count(*)::int` }).from(concepts),
+      db.select({ n: sql<number>`count(distinct ${videoTopics.topic})::int` }).from(videoTopics),
     ]);
     return { videos: v.n, players: p.n, concepts: c.n, topics: t.n };
   },
@@ -485,7 +485,7 @@ export const dbRepo: AdminRepository = {
   async getPlayers(filter, opts): Promise<PlayerListItem[]> {
     const db = getDb();
     const counts = await db
-      .select({ playerId: videoPlayers.playerId, n: sql<number>`count(*)` })
+      .select({ playerId: videoPlayers.playerId, n: sql<number>`count(*)::int` })
       .from(videoPlayers)
       .groupBy(videoPlayers.playerId);
     const byPlayer = new Map(counts.map((c) => [c.playerId, c.n]));
@@ -540,7 +540,7 @@ export const dbRepo: AdminRepository = {
     const [p] = await db.select().from(players).where(eq(players.id, playerId));
     if (!p) return null;
     const [{ n }] = await db
-      .select({ n: sql<number>`count(*)` })
+      .select({ n: sql<number>`count(*)::int` })
       .from(videoPlayers)
       .where(eq(videoPlayers.playerId, playerId));
 
@@ -742,7 +742,7 @@ export const dbRepo: AdminRepository = {
     // The pipeline itself runs out-of-process (scripts/ingest.ts, and the
     // scheduled job). This records intent and reports current state rather
     // than shelling out from a request handler.
-    const [{ n }] = await getDb().select({ n: sql<number>`count(*)` }).from(videos);
+    const [{ n }] = await getDb().select({ n: sql<number>`count(*)::int` }).from(videos);
     const status: SyncStatus = {
       lastSyncAt: nowIso(),
       state: "ok",
